+++ apps/backend/src/app/api/latest/migration-tests/route.tsx
@@ -1,0 +1,30 @@
+import { createSmartRouteHandler } from "@/route-handlers/smart-route-handler";
+import { yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
+import { deindent } from "@stackframe/stack-shared/dist/utils/strings";
+
+export const GET = createSmartRouteHandler({
+  metadata: {
+    hidden: true,
+  },
+  request: yupObject({
+    url: yupString().defined(),
+  }),
+  response: yupObject({
+    statusCode: yupNumber().oneOf([200]).defined(),
+    bodyType: yupString().oneOf(["text"]).defined(),
+    body: yupString().defined(),
+  }),
+  handler: async (req) => {
+    return {
+      statusCode: 200,
+      bodyType: "text",
+      body: deindent`
+        You are on ${req.url}. Please pick a migration test.
+
+        ${[
+          "./smart-route-handler",
+        ].map((path) => `- ${new URL(path, req.url)}`).join("\n")}
+      `,
+    };
+  },
+});